Removed email login functionality
This was a bug with the ability to login with email, which is something that the team told me that they did not need anymore. They told me that firebase handled all of the login and registration page involving phone number and email, so all I had to do was remove the line of code that allowed the user to sign up for email. They did not have the ability to login with email after signing up, so removing the ability to sign up with email is sufficient.